# AsynCUDA13.Client/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "AsynCUDA13.API v1",
    "version": "v1"
  },
  "paths": {
...
    "/api/Media/upload-media": {
      "post": {
        "tags": [
          "Media"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/Media/download-media": {
      "get": {
        "tags": [
          "Media"
        ],
        "parameters": [
          {
            "name": "idOrName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "png"
            }
          },
          {
            "name": "normalizeAudio",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "float",
              "default": 1
            }
          },
          {
            "name": "pullIfRequired",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "name": "keepBufferWhenPulled",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Media/image-data/{idOrName}": {
      "get": {
        "tags": [
          "Media"
        ],
        "parameters": [
          {
            "name": "idOrName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "png"
            }
          },
          {
            "name": "keepData",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ImageData"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageData"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageData"
                }
              }
            }
          }
        }
      }
    },
    "/api/Media/image-preview/{idOrName}": {
      "get": {
        "tags": [
          "Media"
        ],
        "parameters": [
          {
            "name": "idOrName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxDimenions",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 256
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ImageData"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageData"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageData"
                }
              }
            }
          }
        }
      }
    },
    "/api/Media/audio-data/{idOrName}": {
      "get": {
        "tags": [
          "Media"
        ],
        "parameters": [
          {
            "name": "idOrName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chunkSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "overlap",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "float",
              "default": 0.5
            }
          },
          {
            "name": "keepData",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AudioData"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioData"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioData"
                }
              }
            }
          }
        }
      }
    },
    "/api/Media/audio-preview/{idOrName}": {
      "get": {
        "tags": [
          "Media"
        ],
        "parameters": [
          {
            "name": "idOrName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "width",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 512
            }
          },
          {
            "name": "height",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ImageData"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageData"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageData"
                }
              }
            }
          }
        }
      }
    },
    "/api/Media/delete/{idOrName}": {
      "delete": {
        "tags": [
          "Media"
        ],
        "parameters": [
          {
            "name": "idOrName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AudioData": {
        "required": [
          "info"
        ],
        "type": "object",
        "properties": {
          "info": {
            "$ref": "#/components/schemas/AudioInfo"
          },
          "pointer": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "audioDataFloats": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "float"
            },
            "nullable": true
          },
          "audioDataFloatChunks": {
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "type": "number",
                "format": "float"
              }
            },
            "nullable": true
          },
          "chunkSize": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "dataSizeMb": {
            "type": "number",
            "format": "float",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "AudioInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "mediaType": {
            "type": "string",
            "nullable": true
          },
          "sampleRate": {
            "type": "integer",
            "format": "int32"
          },
          "channels": {
            "type": "integer",
            "format": "int32"
          },
          "bitDepth": {
            "type": "integer",
            "format": "int32"
          },
          "length": {
            "type": "string",
            "nullable": true
          },
          "durationSeconds": {
            "type": "number",
            "format": "float"
          },
          "bpm": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "pointer": {
            "type": "string",
            "nullable": true
          },
          "onGpu": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "ImageData": {
        "required": [
          "info"
        ],
        "type": "object",
        "properties": {
          "info": {
            "$ref": "#/components/schemas/ImageInfo"
          },
          "pointer": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "mimeType": {
            "type": "string",
            "nullable": true
          },
          "base64Data": {
            "type": "string",
            "nullable": true
          },
          "base64Image": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "dataSizeMb": {
            "type": "number",
            "format": "float",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "ImageInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "mediaType": {
            "type": "string",
            "nullable": true
          },
          "width": {
            "type": "integer",
            "format": "int32"
          },
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "channels": {
            "type": "integer",
            "format": "int32"
          },
          "originalSizeMb": {
            "type": "number",
            "format": "float"
          },
          "pointer": {
            "type": "string",
            "nullable": true
          },
          "onGpu": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  },
...
}